Use bpfTestContext fields instead of local variables in single-node tests
Replace local hostW, w[2] variables with s.hostW[0], s.w[0][*] struct
fields in bpf_single_node_test.go for consistency with the rest of the
BPF test suite. Move bpfTestContext and related helpers from bpf_test.go
to bpf_helpers_test.go and remove the now-unused kubernetes import.
